Fixed validation steps and end epoch.
https://www.kaggle.com/artgor/object-detection-with-pytorch-lightning 

''' Please, note that some lines with coco_evaluator are commented out. This is because it doesn't work with numpy 1.18 - only with 1.17. And I can't downgrade the version. I'll try to think what to do with this later. '''

My proposed fix are in the code for both validation_step and validation_epoch_end.

I have tried it on a separate project and it works well.